backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiohttp>=3.9.0
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10
langchain>=0.3.7
langchain-openai>=0.2.10
langchain-community>=0.3.5
openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.0.0
python-magic>=0.4.27
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0 